requirements.txt
torch
torchvision
h5py
omegaconf
hydra-core
tqdm
plyfile
ninja
wandb
trimesh
Cython
matplotlib
open3d